Hill Climbing using Exploitation
The main goal is to maximize the number of 1 bits in the range of numbers between 0 and 15. 
First i choose a random number, than i check the fitness, get their neighbours.
Then i exploit each neighbour flipping 1 bit, for example 0101 to 0001, and checking the fitting and keeping the best ones. 
Then i compare these neighbours with my best solution, if the neighbour is better, now the neighbour is my new best solution, else i discard it.
Iterate some times and i always get the global best.